fix: use int16_t for values passed to multS16x16_to32
Update processing.cpp to use int16_t for filtV_div4 and filtI_div4
variables to match the strict type checking in multS16x16_to32().

+++ Mk2_3phase_RFdatalog_temp/processing.cpp
@@ -442,8 +442,8 @@ void processCurrentRawSample(const uint8_t phase, const int16_t rawSample)
   sampleIminusDC += (lpf_gain * lpf_long[phase]);
 
   // calculate the "real power" in this sample pair and add to the accumulated sum
-  const int32_t filtV_div4 = l_sampleVminusDC[phase] >> 2;  // reduce to 16-bits (now x64, or 2^6)
-  const int32_t filtI_div4 = sampleIminusDC >> 2;           // reduce to 16-bits (now x64, or 2^6)
+  const int16_t filtV_div4 = l_sampleVminusDC[phase] >> 2;  // reduce to 16-bits (now x64, or 2^6)
+  const int16_t filtI_div4 = sampleIminusDC >> 2;           // reduce to 16-bits (now x64, or 2^6)
   int32_t instP;
   multS16x16_to32(instP, filtV_div4, filtI_div4);  // 32-bits (now x4096, or 2^12)
   instP >>= 12;                                    // scaling is now x1, as for Mk2 (V_ADC x I_ADC)
@@ -505,7 +505,7 @@ void confirmPolarity(const uint8_t phase)
 void processVoltage(const uint8_t phase)
 {
   // for the Vrms calculation (for datalogging only)
-  const int32_t filtV_div4{ l_sampleVminusDC[phase] >> 2 };  // reduce to 16-bits (now x64, or 2^6)
+  const int16_t filtV_div4{ static_cast< int16_t >(l_sampleVminusDC[phase] >> 2) };  // reduce to 16-bits (now x64, or 2^6)
   int32_t inst_Vsquared;
   multS16x16_to32(inst_Vsquared, filtV_div4, filtV_div4);  // 32-bits (now x4096, or 2^12)
